Move qualify to be a gapic-showcase subcommand
This creates a package `qualifier`
