+++ cmd/gapic-showcase/qualifier/acceptance_suite/passing/echo.proto
