+++ cmd/gapic-showcase/qualifier/acceptance_suite/passing/foo/bar/include.wdREAD
