+++ cmd/gapic-showcase/qualifier/acceptance_suite/passing/include.vcREAD
