+++ cmd/gapic-showcase/qualifier/acceptance_suite/passing/sample.yaml
