+++ cmd/gapic-showcase/qualifier/generator.go
@@ -1,4 +1,4 @@
-package main
+package qualifier
 
 import (
 	"errors"
