+++ cmd/gapic-showcase/qualifier/scenario.go
@@ -1,4 +1,4 @@
-package main
+package qualifier
 
 import (
 	"bytes"
